+++ src/StandardIteratorInterface.php
@@ -107,6 +107,12 @@ interface StandardIteratorInterface extends IteratorAggregate, Countable
     public function tap(callable $callback);
 
     /**
+     * @param callable $callback
+     * @return static
+     */
+    public function sort(callable $callback);
+
+    /**
      * @return bool
      */
     public function isEmpty();