+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Added
 - New collection methods `implode`, `fill` and `fillObject`.
-- New standard iterator methods `take` and `tap`.
+- New standard iterator methods `take`, `tap` and `sort`.
 - Collection can now be json encoded (implements `JsonSerializable`).
 - Collection can now be casted to a string.